tidy up deno_core modules (#5923)
