+++ core/core_isolate.rs
@@ -7,13 +7,14 @@
 
 use rusty_v8 as v8;
 
-use crate::any_error::ErrBox;
 use crate::bindings;
-use crate::js_errors::JSError;
 use crate::ops::*;
 use crate::shared_queue::SharedQueue;
 use crate::shared_queue::RECOMMENDED_SIZE;
+use crate::ErrBox;
+use crate::JSError;
 use crate::ResourceTable;
+use crate::ZeroCopyBuf;
 use futures::future::FutureExt;
 use futures::stream::select;
 use futures::stream::FuturesUnordered;
@@ -24,10 +25,7 @@ use libc::c_void;
 use std::cell::RefCell;
 use std::collections::HashMap;
 use std::convert::From;
-use std::error::Error;
-use std::fmt;
 use std::mem::forget;
-use std::ops::{Deref, DerefMut};
 use std::option::Option;
 use std::pin::Pin;
 use std::rc::Rc;
@@ -36,69 +34,6 @@ use std::task::Context;
 use std::task::Poll;
 
 type PendingOpFuture = Pin<Box<dyn Future<Output = (OpId, Buf)>>>;
-
-/// A ZeroCopyBuf encapsulates a slice that's been borrowed from a JavaScript
-/// ArrayBuffer object. JavaScript objects can normally be garbage collected,
-/// but the existence of a ZeroCopyBuf inhibits this until it is dropped. It
-/// behaves much like an Arc<[u8]>, although a ZeroCopyBuf currently can't be
-/// cloned.
-pub struct ZeroCopyBuf {
-  backing_store: v8::SharedRef<v8::BackingStore>,
-  byte_offset: usize,
-  byte_length: usize,
-}
-
-unsafe impl Send for ZeroCopyBuf {}
-
-impl ZeroCopyBuf {
-  pub fn new(view: v8::Local<v8::ArrayBufferView>) -> Self {
-    let backing_store = view.buffer().unwrap().get_backing_store();
-    let byte_offset = view.byte_offset();
-    let byte_length = view.byte_length();
-    Self {
-      backing_store,
-      byte_offset,
-      byte_length,
-    }
-  }
-}
-
-impl Deref for ZeroCopyBuf {
-  type Target = [u8];
-  fn deref(&self) -> &[u8] {
-    unsafe {
-      bindings::get_backing_store_slice(
-        &self.backing_store,
-        self.byte_offset,
-        self.byte_length,
-      )
-    }
-  }
-}
-
-impl DerefMut for ZeroCopyBuf {
-  fn deref_mut(&mut self) -> &mut [u8] {
-    unsafe {
-      bindings::get_backing_store_slice_mut(
-        &self.backing_store,
-        self.byte_offset,
-        self.byte_length,
-      )
-    }
-  }
-}
-
-impl AsRef<[u8]> for ZeroCopyBuf {
-  fn as_ref(&self) -> &[u8] {
-    &*self
-  }
-}
-
-impl AsMut<[u8]> for ZeroCopyBuf {
-  fn as_mut(&mut self) -> &mut [u8] {
-    &mut *self
-  }
-}
 
 /// Stores a script used to initialize a Isolate
 pub struct Script<'a> {
@@ -625,14 +560,6 @@ fn drain_macrotasks<'s>(
   }
 
   Ok(())
-}
-
-pub(crate) fn attach_handle_to_error(
-  scope: &mut impl v8::InIsolate,
-  err: ErrBox,
-  handle: v8::Local<v8::Value>,
-) -> ErrBox {
-  ErrWithV8Handle::new(scope, err, handle).into()
 }
 
 pub(crate) fn exception_to_err_result<'s, T>(
@@ -1181,44 +1108,5 @@ pub mod tests {
     let startup_data = StartupData::Snapshot(Snapshot::Boxed(snapshot));
     let mut isolate2 = CoreIsolate::new(startup_data, false);
     js_check(isolate2.execute("check.js", "if (a != 3) throw Error('x')"));
-  }
-}
-
-// TODO(piscisaureus): rusty_v8 should implement the Error trait on
-// values of type v8::Global<T>.
-pub struct ErrWithV8Handle {
-  err: ErrBox,
-  handle: v8::Global<v8::Value>,
-}
-
-impl ErrWithV8Handle {
-  pub fn new(
-    scope: &mut impl v8::InIsolate,
-    err: ErrBox,
-    handle: v8::Local<v8::Value>,
-  ) -> Self {
-    let handle = v8::Global::new_from(scope, handle);
-    Self { err, handle }
-  }
-
-  pub fn get_handle(&self) -> &v8::Global<v8::Value> {
-    &self.handle
-  }
-}
-
-unsafe impl Send for ErrWithV8Handle {}
-unsafe impl Sync for ErrWithV8Handle {}
-
-impl Error for ErrWithV8Handle {}
-
-impl fmt::Display for ErrWithV8Handle {
-  fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-    self.err.fmt(f)
-  }
-}
-
-impl fmt::Debug for ErrWithV8Handle {
-  fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-    self.err.fmt(f)
   }
 }
