+++ core/lib.rs
@@ -8,30 +8,44 @@ extern crate lazy_static;
 #[macro_use]
 extern crate log;
 
-mod any_error;
 mod bindings;
+mod core_isolate;
+mod errors;
 mod es_isolate;
 mod flags;
-mod isolate;
-mod js_errors;
 mod module_specifier;
 mod modules;
 mod ops;
 pub mod plugin_api;
 mod resources;
 mod shared_queue;
+mod zero_copy_buf;
 
 pub use rusty_v8 as v8;
 
-pub use crate::any_error::*;
-pub use crate::es_isolate::*;
+pub use crate::core_isolate::js_check;
+pub use crate::core_isolate::CoreIsolate;
+pub use crate::core_isolate::Script;
+pub use crate::core_isolate::Snapshot;
+pub use crate::core_isolate::StartupData;
+pub use crate::errors::ErrBox;
+pub use crate::errors::JSError;
+pub use crate::es_isolate::EsIsolate;
 pub use crate::flags::v8_set_flags;
-pub use crate::isolate::*;
-pub use crate::js_errors::*;
-pub use crate::module_specifier::*;
-pub use crate::modules::*;
-pub use crate::ops::*;
-pub use crate::resources::*;
+pub use crate::module_specifier::ModuleResolutionError;
+pub use crate::module_specifier::ModuleSpecifier;
+pub use crate::modules::ModuleId;
+pub use crate::modules::ModuleLoadId;
+pub use crate::modules::ModuleLoader;
+pub use crate::modules::ModuleSource;
+pub use crate::modules::ModuleSourceFuture;
+pub use crate::modules::RecursiveModuleLoad;
+pub use crate::ops::Buf;
+pub use crate::ops::Op;
+pub use crate::ops::OpAsyncFuture;
+pub use crate::ops::OpId;
+pub use crate::resources::ResourceTable;
+pub use crate::zero_copy_buf::ZeroCopyBuf;
 
 pub fn v8_version() -> &'static str {
   v8::V8::get_version()
