+++ core/modules.rs
@@ -2,10 +2,8 @@
 
 use rusty_v8 as v8;
 
-use crate::any_error::ErrBox;
-use crate::es_isolate::ModuleId;
-use crate::es_isolate::ModuleLoadId;
 use crate::module_specifier::ModuleSpecifier;
+use crate::ErrBox;
 use futures::future::FutureExt;
 use futures::stream::FuturesUnordered;
 use futures::stream::Stream;
@@ -24,6 +22,9 @@ use std::task::Poll;
 lazy_static! {
   pub static ref NEXT_LOAD_ID: AtomicI32 = AtomicI32::new(0);
 }
+
+pub type ModuleId = i32;
+pub type ModuleLoadId = i32;
 
 /// EsModule source code that will be loaded into V8.
 ///
@@ -548,13 +549,20 @@ macro_rules! include_crate_modules {
 mod tests {
   use super::*;
   use crate::es_isolate::EsIsolate;
-  use crate::isolate::js_check;
+  use crate::js_check;
+  use crate::StartupData;
   use futures::future::FutureExt;
   use std::error::Error;
   use std::fmt;
   use std::future::Future;
   use std::sync::Arc;
   use std::sync::Mutex;
+
+  // TODO(ry) Sadly FuturesUnordered requires the current task to be set. So
+  // even though we are only using poll() in these tests and not Tokio, we must
+  // nevertheless run it in the tokio executor. Ideally run_in_task can be
+  // removed in the future.
+  use crate::core_isolate::tests::run_in_task;
 
   struct MockLoader {
     pub loads: Arc<Mutex<Vec<String>>>,
@@ -715,13 +723,6 @@ mod tests {
     if (import.meta.main) throw Error();
     if (import.meta.url != 'file:///d.js') throw Error();
   "#;
-
-  // TODO(ry) Sadly FuturesUnordered requires the current task to be set. So
-  // even though we are only using poll() in these tests and not Tokio, we must
-  // nevertheless run it in the tokio executor. Ideally run_in_task can be
-  // removed in the future.
-  use crate::isolate::tests::run_in_task;
-  use crate::isolate::StartupData;
 
   #[test]
   fn test_recursive_load() {
